자재 마스터 등록시 null exception 발생 수정

+++ src/main/java/com/dksys/biz/user/bm/bm05/service/impl/BM05SvcImpl.java
@@ -110,22 +110,26 @@ public class BM05SvcImpl implements BM05Svc {
 		param.put("comonCd", paramMap.get("comonCd"));  //프로트엔드에 넘어온 화일 저장 위치 정보
 		
 		List<Map<String, String>> uploadFileList = gsonDtl.fromJson(paramMap.get("uploadFileArr"), dtlMap);
-		if (uploadFileList.size() > 0) {
+		if (uploadFileList != null && uploadFileList.size() > 0) {
 			//접근 권한 없으면 Exception 발생 (jobType, userId, comonCd 3개 필수값 필요)
 			param.put("jobType", "fileUp");
 			cm15Svc.selectFileAuthCheck(param);
 		}
 		
 		String[] deleteFileArr = gsonDtl.fromJson(paramMap.get("deleteFileArr"), String[].class);
-		List<String> deleteFileList = Arrays.asList(deleteFileArr);
+		List<String> deleteFileList = (deleteFileArr != null) ? Arrays.asList(deleteFileArr) : new ArrayList<String>();
 		
-		for(String fileKey : deleteFileList) {
-			// 삭제할 파일 하나씩 점검 필요(전체 목록에서 삭제 선택시 필요함)
-			Map<String, String> fileInfo = cm08Svc.selectFileInfo(fileKey);
-			//접근 권한 없으면 Exception 발생
-			param.put("comonCd", fileInfo.get("comonCd"));  //삭제할 파일이 보관된 저장 위치 정보
-			param.put("jobType", "fileDelete");
-			cm15Svc.selectFileAuthCheck(param);
+		if (deleteFileList.size() > 0) {
+			for(String fileKey : deleteFileList) {
+				// 삭제할 파일 하나씩 점검 필요(전체 목록에서 삭제 선택시 필요함)
+				Map<String, String> fileInfo = cm08Svc.selectFileInfo(fileKey);
+				if (fileInfo != null) {
+					//접근 권한 없으면 Exception 발생
+					param.put("comonCd", fileInfo.get("comonCd"));  //삭제할 파일이 보관된 저장 위치 정보
+					param.put("jobType", "fileDelete");
+					cm15Svc.selectFileAuthCheck(param);
+				}
+			}
 		}
 		//---------------------------------------------------------------
 		//첨부 화일 권한체크  끝
@@ -136,15 +140,17 @@ public class BM05SvcImpl implements BM05Svc {
 		//---------------------------------------------------------------
 		//첨부 화일 처리 시작
 		//---------------------------------------------------------------
-		if (uploadFileList.size() > 0) {
+		if (uploadFileList != null && uploadFileList.size() > 0) {
 			paramMap.put("fileTrgtTyp", paramMap.get("pgmId"));
 			paramMap.put("fileTrgtKey", paramMap.get("fileTrgtKey"));
 			paramMap.put("coCd", "TRN");
 			cm08Svc.uploadFile(paramMap, mRequest);
 		}
 		
-		for(String fileKey : deleteFileList) {
-			cm08Svc.deleteFile(fileKey);
+		if (deleteFileList.size() > 0) {
+			for(String fileKey : deleteFileList) {
+				cm08Svc.deleteFile(fileKey);
+			}
 		}
 		//---------------------------------------------------------------
 		//첨부 화일 처리  끝
@@ -162,14 +168,15 @@ public class BM05SvcImpl implements BM05Svc {
 		//첨부 화일 처리 권한체크 시작 -->파일 업로드, 삭제 권한 없으면 Exception 처리 됨
 		//   필수값 :  jobType, userId, comonCd
 		//---------------------------------------------------------------
+		HashMap<String, String> param = new HashMap<>();
+		param.put("userId", paramMap.get("userId"));
+		param.put("comonCd", paramMap.get("comonCd"));
+		
 		List<Map<String, String>> uploadFileList = gsonDtl.fromJson(paramMap.get("uploadFileArr"), dtlMap);
-		if (uploadFileList.size() > 0) {
-			HashMap<String, String> param = new HashMap<>();
-			//접근 권한 없으면 Exception 발생
-			param.putAll(paramMap);
+		if (uploadFileList != null && uploadFileList.size() > 0) {
 			param.put("jobType", "fileUp");
-			param.put("coCd", "TRN");
-			cm15Svc.selectFileAuthCheck(paramMap);
+			param.put("coCd", paramMap.get("coCd") != null ? paramMap.get("coCd") : "TRN");
+			cm15Svc.selectFileAuthCheck(param);
 		}
 		//---------------------------------------------------------------
 		//첨부 화일 권한체크  끝
@@ -182,10 +189,9 @@ public class BM05SvcImpl implements BM05Svc {
 		//---------------------------------------------------------------
 		//첨부 화일 처리 시작  (처음 등록시에는 화일 삭제할게 없음)
 		//---------------------------------------------------------------
-		if (uploadFileList.size() > 0) {
+		if (uploadFileList != null && uploadFileList.size() > 0) {
 			paramMap.put("fileTrgtTyp", paramMap.get("pgmId"));
 			paramMap.put("fileTrgtKey", paramMap.get("fileTrgtKey"));
-			paramMap.put("coCd", "TRN");
 			cm08Svc.uploadFile(paramMap, mRequest);
 		}
 		//---------------------------------------------------------------
@@ -204,13 +210,16 @@ public class BM05SvcImpl implements BM05Svc {
 		List<Map<String, String>> deleteFileList = cm08Svc.selectFileListAll(paramMap);
 		HashMap<String, String> param = new HashMap<>();
 		param.put("jobType", "fileDelete");
-		param.put("coCd", "TRN");
+		param.put("coCd", paramMap.get("coCd") != null ? paramMap.get("coCd") : "TRN");
 		param.put("userId", paramMap.get("userId"));
-		if (deleteFileList.size() > 0) {
+		
+		if (deleteFileList != null && deleteFileList.size() > 0) {
 			for (Map<String, String> dtl : deleteFileList) {
 				//접근 권한 없으면 Exception 발생
-				param.put("comonCd",  dtl.get("comonCd"));
-				cm15Svc.selectFileAuthCheck(param);
+				if (dtl.get("comonCd") != null) {
+					param.put("comonCd", dtl.get("comonCd"));
+					cm15Svc.selectFileAuthCheck(param);
+				}
 			}
 		}
 		//---------------------------------------------------------------
@@ -222,10 +231,12 @@ public class BM05SvcImpl implements BM05Svc {
 		//---------------------------------------------------------------
 		//첨부 화일 처리 시작  (처음 등록시에는 화일 삭제할게 없음)
 		//---------------------------------------------------------------
-		if (deleteFileList.size() > 0) {
+		if (deleteFileList != null && deleteFileList.size() > 0) {
 			for (Map<String, String> deleteDtl : deleteFileList) {
-				String fileKey = deleteDtl.get("fileKey").toString();
-				cm08Svc.deleteFile( fileKey );
+				if (deleteDtl.get("fileKey") != null) {
+					String fileKey = String.valueOf(deleteDtl.get("fileKey"));
+					cm08Svc.deleteFile(fileKey);
+				}
 			}
 		}
 		//---------------------------------------------------------------
